+++ packages/suite/src/components/suite/AppNavigation/index.tsx
@@ -1,19 +1,23 @@
-import React from 'react';
+import React, { useRef, useLayoutEffect, useState } from 'react';
 import styled from 'styled-components';
-import { Icon, variables, IconProps, useTheme } from '@trezor/components';
+import { variables, IconProps, useTheme, Button, Icon, Dropdown } from '@trezor/components';
 import { Translation } from '@suite-components';
 import { useSelector, useActions } from '@suite-hooks';
 import * as routerActions from '@suite-actions/routerActions';
 import { Route } from '@suite-types';
+import useResizeObserver from '@react-hook/resize-observer';
 
 const { FONT_WEIGHT, FONT_SIZE } = variables;
+
+const SECONDARY_MENU_BUTTON_MARGIN = '12px';
 
 const Wrapper = styled.div`
     display: flex;
     width: 100%;
-    border-top: 1px solid ${props => props.theme.STROKE_GREY};
-    overflow-x: auto;
+    /* overflow-x: auto; */
     scrollbar-width: none; /* Firefox */
+    justify-content: space-between;
+    position: relative;
 
     &::-webkit-scrollbar {
         /* WebKit */
@@ -22,107 +26,252 @@ const Wrapper = styled.div`
     }
 `;
 
-const StyledNavLink = styled.div<{ isComingSoon?: boolean; active?: boolean }>`
-    cursor: ${props => (props.isComingSoon ? 'default' : 'pointer')};
+const Primary = styled.div`
+    display: flex;
+`;
+
+const Secondary = styled.div`
+    display: flex;
+    align-items: center;
+`;
+
+const SecondaryMenu = styled.div<{ visible: boolean }>`
+    display: flex;
+    align-items: center;
+    ${props => !props.visible && `opacity: 0;`}
+    & > * + * {
+        margin-left: ${SECONDARY_MENU_BUTTON_MARGIN};
+    }
+`;
+
+const SecondaryMenuCondensed = styled.div`
+    position: absolute;
+    top: 0;
+    right: 0;
+    z-index: 3;
+    background: ${props => props.theme.BG_WHITE};
+    background: linear-gradient(
+        90deg,
+        transparent 0%,
+        ${props => props.theme.BG_WHITE} 10%,
+        ${props => props.theme.BG_WHITE} 100%
+    );
+    height: 100%;
+    width: 70px;
+    margin: 0 -33px 0;
+    padding: 0 0 0 13px;
+`;
+
+const StyledNavLink = styled.div<{ active?: boolean }>`
+    cursor: pointer;
     font-size: ${FONT_SIZE.NORMAL};
     color: ${props =>
-        props.active ? props => props.theme.TYPE_GREEN : props => props.theme.TYPE_LIGHT_GREY};
+        props.active ? props => props.theme.TYPE_DARK_GREY : props => props.theme.TYPE_LIGHT_GREY};
     font-weight: ${FONT_WEIGHT.MEDIUM};
     display: flex;
     align-items: center;
-    padding: 14px 8px 12px 6px;
+    padding: 23px 0;
     white-space: nowrap;
     border-bottom: 2px solid
-        ${props => (props.active ? props => props.theme.BG_GREEN : 'transparent')};
-
-    & + & {
-        margin-left: 42px;
+        ${props => (props.active ? props => props.theme.TYPE_DARK_GREY : 'transparent')};
+    margin-right: 40px;
+    &:last-child {
+        margin-right: ${SECONDARY_MENU_BUTTON_MARGIN};
     }
 `;
 
 const IconWrapper = styled.div`
+    margin-right: 10px;
+`;
+
+const StyledBackLink = styled.div`
+    cursor: pointer;
+    font-size: ${FONT_SIZE.SMALL};
+    color: ${props => props.theme.TYPE_DARK_GREY};
+    font-weight: ${FONT_WEIGHT.MEDIUM};
     display: flex;
-    justify-content: center;
-    width: 16px;
+    align-items: center;
+    padding: 23px 0 25px;
+    white-space: nowrap;
 `;
 
 const Text = styled.div`
-    padding-left: 8px;
     position: relative;
 `;
 
-const Soon = styled.div`
-    position: absolute;
-    top: -10px;
-    right: -15px;
-    background: #e1f2dc;
-    font-weight: ${variables.FONT_WEIGHT.BOLD};
-    padding: 2px 4px;
-    color: ${props => props.theme.TYPE_GREEN};
-    border-radius: 4px;
-    font-size: 9px;
+const StyledIcon = styled(Icon)`
+    margin-right: 10px;
 `;
 
-interface Props {
-    items: {
-        route: Route['name'];
-        title: JSX.Element;
-        icon: IconProps['icon'];
-        isComingSoon?: boolean;
-        'data-test'?: string;
-    }[];
-}
-
-const isRouteActive = (routeName?: Route['name'], route?: Route['name']): boolean => {
-    // coinmarket has multiple routes, match them all
-    if (routeName?.startsWith('wallet-coinmarket') && route?.startsWith('wallet-coinmarket')) {
-        return true;
-    }
-    return routeName === route;
+export type AppNavigationItem = {
+    id: string;
+    callback: () => void;
+    title: JSX.Element;
+    position: 'primary' | 'secondary';
+    icon?: IconProps['icon'];
+    'data-test'?: string;
+    isHidden?: () => boolean;
 };
 
-const AppNavigation = ({ items }: Props) => {
+interface Props {
+    items: AppNavigationItem[];
+    primaryContent?: React.ReactNode;
+}
+interface MenuWidths {
+    primary: number;
+    secondary: number;
+    wrapper: number;
+}
+
+const isRouteActive = (routeName?: Route['name'], id?: string): boolean => {
+    return routeName === id;
+};
+
+const isSubsection = (routeName: Route['name']): boolean => {
+    return !(
+        routeName.startsWith('settings') ||
+        routeName === 'wallet-index' ||
+        routeName === 'wallet-details'
+    );
+};
+
+const isSecondaryMenuOverflown = ({ primary, secondary, wrapper }: MenuWidths) =>
+    primary + secondary >= wrapper;
+
+const AppNavigation = ({ items, primaryContent }: Props) => {
     const theme = useTheme();
-    const { routeName, params } = useSelector(state => ({
+    const [condensedSecondaryMenuVisible, setCondensedSecondaryMenuVisible] = useState<boolean>(
+        false,
+    );
+    const wrapper = useRef<HTMLDivElement>(null);
+    const primary = useRef<HTMLDivElement>(null);
+    const secondary = useRef<HTMLDivElement>(null);
+    const { routeName } = useSelector(state => ({
         routeName: state.router.route?.name,
-        params: state.wallet.selectedAccount.params,
     }));
     const { goto } = useActions({
         goto: routerActions.goto,
     });
 
-    return (
-        <Wrapper>
-            {items.map(item => {
-                const { route, title, icon, isComingSoon, ...restItemProps } = item;
-                const active = isRouteActive(routeName, route);
-                return (
-                    <StyledNavLink
-                        key={route}
-                        active={active}
-                        isComingSoon={isComingSoon}
-                        onClick={isComingSoon ? undefined : () => goto(route, params, !params)}
-                        {...restItemProps}
-                    >
-                        <IconWrapper>
-                            <Icon
-                                size={18}
-                                icon={icon}
-                                color={active ? theme.TYPE_GREEN : undefined}
-                            />
-                        </IconWrapper>
+    useLayoutEffect(() => {
+        if (primary.current && secondary.current && wrapper.current) {
+            setCondensedSecondaryMenuVisible(
+                isSecondaryMenuOverflown({
+                    primary: primary.current.getBoundingClientRect().width,
+                    secondary: secondary.current.getBoundingClientRect().width,
+                    wrapper: wrapper.current.getBoundingClientRect().width,
+                }),
+            );
+        }
+    }, [wrapper, primary, secondary]);
+    useResizeObserver(wrapper, entry => {
+        if (primary.current && secondary.current && wrapper.current) {
+            setCondensedSecondaryMenuVisible(
+                isSecondaryMenuOverflown({
+                    primary: primary.current.getBoundingClientRect().width,
+                    secondary: secondary.current.getBoundingClientRect().width,
+                    wrapper: entry.contentRect.width,
+                }),
+            );
+        }
+    });
 
-                        <Text>
-                            {isComingSoon && (
-                                <Soon>
-                                    <Translation id="TR_NAV_EXCHANGE_SOON" />
-                                </Soon>
-                            )}
-                            {title}
-                        </Text>
-                    </StyledNavLink>
-                );
-            })}
+    return (
+        <Wrapper ref={wrapper}>
+            {routeName && isSubsection(routeName) ? (
+                <Primary>
+                    <StyledBackLink onClick={() => goto('wallet-index', undefined, true)}>
+                        <StyledIcon icon="ARROW_LEFT" size={16} />
+                        <Translation id="TR_BACK" />
+                    </StyledBackLink>
+                </Primary>
+            ) : (
+                <>
+                    <Primary ref={primary}>
+                        {primaryContent ||
+                            items
+                                .filter(item => item.position === 'primary')
+                                .map(item => {
+                                    const { id, title } = item;
+                                    const active = isRouteActive(routeName, id);
+                                    return (
+                                        <StyledNavLink
+                                            key={id}
+                                            active={active}
+                                            onClick={item.callback}
+                                            {...(item['data-test'] && {
+                                                'data-test': item['data-test'],
+                                            })}
+                                        >
+                                            {item.icon && (
+                                                <IconWrapper>
+                                                    <Icon
+                                                        size={18}
+                                                        icon={item.icon}
+                                                        color={
+                                                            active
+                                                                ? theme.TYPE_DARK_GREY
+                                                                : undefined
+                                                        }
+                                                    />
+                                                </IconWrapper>
+                                            )}
+
+                                            <Text>{title}</Text>
+                                        </StyledNavLink>
+                                    );
+                                })}
+                    </Primary>
+                    <Secondary ref={secondary}>
+                        {condensedSecondaryMenuVisible && (
+                            <SecondaryMenuCondensed>
+                                <Dropdown
+                                    alignMenu="right"
+                                    offset={8}
+                                    items={[
+                                        {
+                                            key: 'group1',
+                                            options: items
+                                                .filter(item => item.position === 'secondary')
+                                                .map(item => {
+                                                    const { id, title } = item;
+                                                    return {
+                                                        key: id,
+                                                        callback: () => {
+                                                            item.callback();
+                                                            return true;
+                                                        },
+                                                        label: title,
+                                                    };
+                                                }),
+                                        },
+                                    ]}
+                                />
+                            </SecondaryMenuCondensed>
+                        )}
+                        <SecondaryMenu visible={!condensedSecondaryMenuVisible}>
+                            {items
+                                .filter(item => item.position === 'secondary')
+                                .map(item => {
+                                    const { id, title } = item;
+                                    return (
+                                        <Button
+                                            key={id}
+                                            variant="secondary"
+                                            onClick={item.callback}
+                                            {...(item['data-test'] && {
+                                                'data-test': item['data-test'],
+                                            })}
+                                            isDisabled={condensedSecondaryMenuVisible}
+                                        >
+                                            <Text>{title}</Text>
+                                        </Button>
+                                    );
+                                })}
+                        </SecondaryMenu>
+                    </Secondary>
+                </>
+            )}
         </Wrapper>
     );
 };