+++ packages/suite/src/components/suite/AppNavigationPanel/index.tsx
@@ -1,5 +1,6 @@
 import styled from 'styled-components';
 import { H1, variables } from '@trezor/components';
+import { useInView } from 'react-intersection-observer';
 import React from 'react';
 import { MAX_WIDTH, MAX_WIDTH_WALLET_CONTENT } from '@suite-constants/layout';
 
@@ -27,7 +28,8 @@ const Content = styled.div<Pick<Props, 'maxWidth'>>`
 const BasicInfo = styled.div`
     display: flex;
     flex-direction: column;
-    margin-bottom: 20px;
+    padding-bottom: 20px;
+    border-bottom: 1px solid ${props => props.theme.STROKE_GREY};
 `;
 
 const Title = styled(H1)`
@@ -36,6 +38,13 @@ const Title = styled(H1)`
     color: ${props => props.theme.TYPE_DARK_GREY};
     white-space: nowrap;
     overflow: hidden;
+`;
+
+const Aside = styled.div`
+    display: flex;
+    & > * + * {
+        margin-left: 10px;
+    }
 `;
 
 const Row = styled.div`
@@ -47,31 +56,85 @@ const Row = styled.div`
 
 const TitleRow = styled(Row)`
     margin-bottom: 6px;
-    align-items: normal;
+`;
+
+const STICKY_MENU_HEIGHT = 70;
+const StickyMenu = styled.div<{ visible: boolean }>`
+    position: sticky;
+    top: 0;
+    left: 0;
+    height: 0;
+    width: 100%;
+    z-index: 2;
+    transform: translate(0, -${STICKY_MENU_HEIGHT + 1}px);
+    transition: all 0.3s ease;
+
+    ${props =>
+        props.visible &&
+        `
+        transform: translate(0, 0);
+        `}
+
+    @media screen and (max-width: ${variables.SCREEN_SIZE.SM}) {
+        display: none;
+    }
+`;
+
+const StickyMenuInner = styled.div`
+    background: ${props => props.theme.BG_WHITE};
+    border-bottom: 1px solid ${props => props.theme.STROKE_GREY};
+`;
+
+const StickyMenuHolder = styled.div<Pick<Props, 'maxWidth'>>`
+    height: ${STICKY_MENU_HEIGHT}px;
+    max-width: ${props => (props.maxWidth === 'default' ? MAX_WIDTH : MAX_WIDTH_WALLET_CONTENT)};
+    align-items: center;
+    margin: 0 auto;
 `;
 
 interface Props {
     title: React.ReactNode;
+    ticker?: React.ReactNode;
     dropdown?: React.ReactNode;
     maxWidth: 'small' | 'default';
     children?: React.ReactNode;
     navigation?: React.ReactNode;
+    navigationSticky?: React.ReactNode;
 }
 
 const AppNavigationPanel = (props: Props) => {
+    const { ref, inView } = useInView({
+        delay: 100,
+        initialInView: true,
+    });
+
     return (
-        <Wrapper>
-            <Content maxWidth={props.maxWidth}>
-                <BasicInfo>
-                    <TitleRow>
-                        <Title noMargin>{props.title}</Title>
-                        {props.dropdown && props.dropdown}
-                    </TitleRow>
-                    {props.children && <Row>{props.children}</Row>}
-                </BasicInfo>
-                {props.navigation}
-            </Content>
-        </Wrapper>
+        <>
+            {props.navigationSticky && (
+                <StickyMenu visible={!inView}>
+                    <StickyMenuInner>
+                        <StickyMenuHolder maxWidth={props.maxWidth}>
+                            {props.navigationSticky}
+                        </StickyMenuHolder>
+                    </StickyMenuInner>
+                </StickyMenu>
+            )}
+            <Wrapper ref={ref}>
+                <Content maxWidth={props.maxWidth}>
+                    <BasicInfo>
+                        <TitleRow>
+                            <Title noMargin>{props.title}</Title>
+                            <Aside>
+                                {props.ticker && props.ticker}
+                                {props.dropdown && props.dropdown}
+                            </Aside>
+                        </TitleRow>
+                        {props.children && <Row>{props.children}</Row>}
+                    </BasicInfo>
+                    {props.navigation}
+                </Content>
+            </Wrapper>
+        </>
     );
 };